started working with sqlite

+++ get_movies_and_cast.py
@@ -12,13 +12,19 @@ session = HTMLSession()
 tmdb.API_KEY = '044b74de638995f58e819e4736f5b29c'
 
 # Need to use another API for better results or a combination of two APIs
-def tmdbSearch(movie):
-    search = tmdb.Search()
-    response = search.movie(query=movie)
-    if (len(search.results) <= 0):
-        return (False)
+def tmdbSearch(year, movie):
+    url = "https://api.themoviedb.org/3/search/movie?api_key=c4d48123a2a6f5d1d47be09f93ecc8db&language=en-US&query={}&page=1&year={}".format(movie, str(year))
+    # search = tmdb.Search()
+    # response = search.movie(query=movie)
+    response = req.get(url)
+    if response.status_code == 200:
+        resp = response.json()
+        if resp["total_results"]>>0:
+            return(True)
+        else:
+            return(False)
     else:
-        return(True)
+        return(False)
     # s = search.results[0]
     # print(s['title'] + " :")
     # movie = tmdb.Movies(s['id'])
@@ -32,6 +38,23 @@ def tmdbSearch(movie):
     #     except Exception:
     #         pass
 
+def webscrapeMovies(country, year):
+    movies = []
+    print('\t\tChecking the movies for the year '+str(year))
+    data = session.get('https://www.the-numbers.com/'+country+'/movies/year/'+str(year))
+    soup = BeautifulSoup(data.content, 'html.parser', from_encoding="utf-8")
+    test = soup.find_all('tr')
+    minimum = min(len(test), 41)
+    print('\t\tExtracting '+str(minimum)+' movie titles')
+    i=1
+    while i<minimum:
+        arr = str(test[i]).split('#tab=summary')[1][2:].split('</a>')[0]
+        movies.append(html.unescape(arr))
+        print(html.unescape(arr))
+        i +=1
+    return(movies)
+
+
 def get_movies(year, country):
 
     dirName = country
@@ -42,29 +65,37 @@ def get_movies(year, country):
         print("Directory " , dirName ,  " already exists")
 
     movies = {}
-    while year >2018 :
-        print('\t\tChecking the movies for the year '+str(year))
-        movies[year]=[]
-        data = session.get('https://www.the-numbers.com/'+country+'/movies/year/'+str(year))
-        soup = BeautifulSoup(data.content, 'html.parser', from_encoding="utf-8")
-        test = soup.find_all('tr')
-        minimum = min(len(test), 41)
-        print('\t\tExtracting '+str(minimum)+' movie titles')
-        i=1
-        while i<minimum:
-            arr = str(test[i]).split('#tab=summary')[1][2:].split('</a>')[0]
-            movies[year].append(html.unescape(arr))
-            print(html.unescape(arr))
-            i +=1
-        filename = country+'/'+str(year)+'.txt'
-
-        for year in movies:
+    while year >1999 :
+        if os.path.isfile('./'+country+'/'+str(year)+'.txt')==False:
+            movies = webscrapeMovies(country, year)
+            filename = country+'/'+str(year)+'.txt'
             with open(filename, "w", encoding='utf-8') as txt_file:
-                for line in movies[year]:
+                for line in movies:
                     txt_file.write(str(line) + "\n")
-        year-=1
-    print(movies)
-    return(movies)
+        else:
+            movies = open('./'+country+'/'+str(year)+'.txt')
+            nb_lines = 0
+            for line in movies:
+                nb_lines += 1
+            if nb_lines<40:
+                answer = input("The number of movies saved for {} is less than 40.\nDo you want to proceed anyways? (y/n)".format(str(year)))
+                if answer.lower() == 'n':
+                    print("well that's sad")
+                elif answer.lower() == 'y':
+                    print("you're super cool")
+                else:
+                    print("can you read???")
+            year-=1
+
+    #
+    #
+    #     for year in movies:
+    #         with open(filename, "w", encoding='utf-8') as txt_file:
+    #             for line in movies[year]:
+    #                 txt_file.write(str(line) + "\n")
+    #     year-=1
+    # print(movies)
+    # return(movies)
 
 def searchMovie(year, movie):
     req_url = "http://www.omdbapi.com/?t="+movie+"&y="+str(year)+"&plot=full&apikey=50f2d223"
@@ -113,8 +144,10 @@ def getmoviesfromthefuckingfiles(year, country):
     while year > 1999:
         input_file = open('./' + country + '/'+ str(year) + '.txt', 'r', encoding='utf_8')
         for line in input_file:
+            if 'Star Wars' in line:
+                line = "Star Wars"
             print('\tSearching movie {}'.format(line.strip()))
-            tmdb_result = tmdbSearch(line)
+            tmdb_result = tmdbSearch(year, line)
             omdb_result = searchMovie(year, line)
             if omdb_result == 'error':
                 print('Not found in omdb')
@@ -126,4 +159,4 @@ def getmoviesfromthefuckingfiles(year, country):
     print(str(len(movies_not_found_tmdb))+" weren't found in tmdb")
     print(str(len(movies_not_found_omdb))+" weren't found in omdb")
     diff = list(set(movies_not_found_omdb) - set(movies_not_found_tmdb))
-    print('they have '+len(str(diff))+' different films not found.')
+    print('they have '+str(len(diff))+' different films not found.')
